Working version of dask delayed, split off entries module
